pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sbai"
version = "0.1.0"
authors = [
    { name="Shayan Bhutto", email="py@shayanbhutto.com" },
]
description = "A short description of the SBAI package"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "openai>=1.0",
    "google-generativeai>=0.3.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/shayanbhutto/sbai"